requirements.txt
# Core web framework
fastapi==0.103.1
uvicorn==0.23.2
aiohttp==3.11.0b0
jinja2==3.1.6

# Uvicorn adapters for better performance
websockets>=11.0.3
httptools>=0.6.0
uvloop>=0.17.0 ; sys_platform != 'win32'  # uvloop doesn't support Windows

# HTTP client
httpx>=0.23.0

# Authentication
python-jose[cryptography]>=1.3.0  # JWT operations
passlib[bcrypt]                   # Password hashing
python-multipart                  # Form data parsing
itsdangerous>=2.1.2               # Session middleware

# OIDC
authlib>=1.2.0                    # OAuth/OIDC library
starlette-session>=0.4.3          # Session management

# Utilities
python-multipart==0.0.18
pydantic==2.4.0
certifi>=2023.7.22
python-dotenv==1.0.0 